Fix some warnings that have been previously ignored(disabled). Add SubWCRev and pre-build instructions to retrieve the revision number. Update ChaosLauncherPlugin with newlines and remove several comments and stuff.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@1382 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/BW/UnitTarget.cpp
@@ -16,7 +16,7 @@ namespace BW
 #pragma warning(disable:4311)
   UnitTarget::UnitTarget(BWAPI::UnitImpl* target)
   {
-    this->targetID = (int)((int)target->getOriginalRawData() - (int)BWDATA_UnitNodeTable) / 336 | 1 << 11;
+    this->targetID = (u16)(((int)target->getOriginalRawData() - (int)BWDATA_UnitNodeTable) / 336 | 1 << 11);
   }
   //-------------------------------------------------- GETTARGET ---------------------------------------------
   u16 UnitTarget::getTarget() const
